requirements.txt
ultralytics
rknn-toolkit2
numpy
opencv-python
torch
torchvision
PyYAML
